+++ spec/valvat/lookup_spec.rb
@@ -27,6 +27,47 @@ describe Valvat::Lookup do
         Valvat::Lookup.validate("").should eql(false)
       end
     end
+    
+    context "with details" do
+      it "returns hash of details instead of true" do
+        Valvat::Lookup.validate("IE6388047V", :detail => true).should eql({
+          :country_code=>"IE", 
+          :vat_number=>"6388047V", 
+          :request_date=> Date.today, 
+          :name=>"GOOGLE IRELAND LIMITED", 
+          :address=>"1ST & 2ND FLOOR ,GORDON HOUSE ,BARROW STREET ,DUBLIN 4"
+        })
+        
+        Valvat::Lookup.validate("LU21416127", :detail => true).should eql({
+          :country_code=>"LU", 
+          :vat_number=>"21416127", 
+          :request_date=> Date.today, 
+          :name=>"EBAY EUROPE S.A R.L.", 
+          :address=>"22, BOULEVARD ROYAL\nL-2449  LUXEMBOURG"
+        })
+      end
+      
+      it "still returns false on not existing vat number" do
+        Valvat::Lookup.validate("BE08173319921", :detail => true).should eql(false)
+      end
+    end
+    
+    context "with request identifier" do
+      it "returns hash of details instead of true" do
+        response = Valvat::Lookup.validate("LU21416127", :requester_vat => "IE6388047V")
+        response[:request_identifier].size.should eql(16)
+        response[:request_identifier] = "some_uniq_string"
+        response.should eql({
+          :country_code=>"LU", 
+          :vat_number=>"21416127", 
+          :request_date=> Date.today, 
+          :name => "EBAY EUROPE S.A R.L.", 
+          :address => "22, BOULEVARD ROYAL\nL-2449  LUXEMBOURG",
+          :company_type => nil,
+          :request_identifier => "some_uniq_string"
+        })
+      end
+    end
 
     # TODO : Reactivate with coorect "down" response
     # context "country web service down" do